requirements.txt
torch
torchvision
hydra-core
scikit-learn
numpy
Pillow
mlflow
optuna
plotly
tifffile